Use uint32_t instead of unsigned int
Had the wrong impression of the minimum guaranteeed size of unsigned int
as being that of unsigned long. Though it didn't factor in the compilers
I've been using thus far. The Fixed-width of uint32_t seems appropriate
for when wanting to do bitwise operations on the addresses like
calculating tags, and indexes to buffers.

Edit: on school linux server must compile with '-std=c++0x' flag